Major code refactor
I really want to give ionar a really really large "hug" right now

+++ src/main/java/me/ionar/salhack/gui/chat/SalGuiTextField.java
@@ -10,33 +10,27 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 import net.minecraft.util.math.MathHelper;
 
-public class SalGuiTextField extends GuiTextField
-{
+public class SalGuiTextField extends GuiTextField {
     public boolean HandleCommandsInternal = true;
 
-    public SalGuiTextField(int componentId, FontRenderer RenderUtilObj, int x, int y, int par5Width, int par6Height)
-    {
+    public SalGuiTextField(int componentId, FontRenderer RenderUtilObj, int x, int y, int par5Width, int par6Height) {
         super(componentId, RenderUtilObj, x, y, par5Width, par6Height);
     }
 
-    public SalGuiTextField(int componentId, FontRenderer RenderUtilObj, int x, int y, int par5Width, int par6Height, boolean b)
-    {
+    public SalGuiTextField(int componentId, FontRenderer RenderUtilObj, int x, int y, int par5Width, int par6Height, boolean b) {
         this(componentId, RenderUtilObj, x, y, par5Width, par6Height);
         HandleCommandsInternal = b;
     }
 
     @Override
-    public void setSelectionPos(int position)
-    {
+    public void setSelectionPos(int position) {
         int i = this.text.length();
 
-        if (position > i)
-        {
+        if (position > i) {
             position = i;
         }
 
-        if (position < 0)
-        {
+        if (position < 0) {
             position = 0;
         }
 
@@ -44,8 +38,7 @@ public class SalGuiTextField extends GuiTextField
 
         //if (this.fontRenderer != null)
         {
-            if (this.lineScrollOffset > i)
-            {
+            if (this.lineScrollOffset > i) {
                 this.lineScrollOffset = i;
             }
 
@@ -53,60 +46,47 @@ public class SalGuiTextField extends GuiTextField
             String s = RenderUtil.trimStringToWidth(this.text.substring(this.lineScrollOffset), j);
             int k = s.length() + this.lineScrollOffset;
 
-            if (position == this.lineScrollOffset)
-            {
+            if (position == this.lineScrollOffset) {
                 this.lineScrollOffset -= RenderUtil.trimStringToWidth(this.text, j, true).length();
             }
 
-            if (position > k)
-            {
+            if (position > k) {
                 this.lineScrollOffset += position - k;
-            }
-            else if (position <= this.lineScrollOffset)
-            {
+            } else if (position <= this.lineScrollOffset) {
                 this.lineScrollOffset -= this.lineScrollOffset - position;
             }
 
             this.lineScrollOffset = MathHelper.clamp(this.lineScrollOffset, 0, i);
         }
     }
-    
+
     @Override
-    public boolean mouseClicked(int mouseX, int mouseY, int mouseButton)
-    {
+    public boolean mouseClicked(int mouseX, int mouseY, int mouseButton) {
         boolean flag = mouseX >= this.x && mouseX < this.x + this.width && mouseY >= this.y && mouseY < this.y + this.height;
 
-        if (this.canLoseFocus)
-        {
+        if (this.canLoseFocus) {
             this.setFocused(flag);
         }
 
-        if (this.isFocused && flag && mouseButton == 0)
-        {
+        if (this.isFocused && flag && mouseButton == 0) {
             int i = mouseX - this.x;
 
-            if (this.enableBackgroundDrawing)
-            {
+            if (this.enableBackgroundDrawing) {
                 i -= 4;
             }
 
             String s = RenderUtil.trimStringToWidth(this.text.substring(this.lineScrollOffset), this.getWidth());
             this.setCursorPosition(RenderUtil.trimStringToWidth(s, i).length() + this.lineScrollOffset);
             return true;
-        }
-        else
-        {
+        } else {
             return false;
         }
     }
 
     @Override
-    public void drawTextBox()
-    {
-        if (this.getVisible())
-        {
-            if (this.getEnableBackgroundDrawing())
-            {
+    public void drawTextBox() {
+        if (this.getVisible()) {
+            if (this.getEnableBackgroundDrawing()) {
                 drawRect(this.x - 1, this.y - 1, this.x + this.width + 1, this.y + this.height + 1, -6250336);
                 drawRect(this.x, this.y, this.x + this.width, this.y + this.height, -16777216);
             }
@@ -121,78 +101,62 @@ public class SalGuiTextField extends GuiTextField
             int i1 = this.enableBackgroundDrawing ? this.y + (this.height - 8) / 2 : this.y;
             int j1 = l;
 
-            if (k > s.length())
-            {
+            if (k > s.length()) {
                 k = s.length();
             }
 
-            if (!s.isEmpty())
-            {
+            if (!s.isEmpty()) {
                 String s1 = flag ? s.substring(0, j) : s;
-                j1 = (int) RenderUtil.drawStringWithShadow(s1, (float)l, (float)i1, 0xFFFFFF)+6;
+                j1 = (int) RenderUtil.drawStringWithShadow(s1, (float) l, (float) i1, 0xFFFFFF) + 6;
             }
 
             boolean flag2 = this.cursorPosition < this.text.length() || this.text.length() >= this.getMaxStringLength();
             int k1 = j1;
 
-            if (!flag)
-            {
+            if (!flag) {
                 k1 = j > 0 ? l + this.width : l;
-            }
-            else if (flag2)
-            {
+            } else if (flag2) {
                 k1 = j1 - 1;
                 --j1;
             }
 
-            if (!s.isEmpty() && flag && j < s.length())
-            {
-                j1 = (int) RenderUtil.drawStringWithShadow(s.substring(j), (float)j1, (float)i1, i);
+            if (!s.isEmpty() && flag && j < s.length()) {
+                j1 = (int) RenderUtil.drawStringWithShadow(s.substring(j), (float) j1, (float) i1, i);
             }
 
-            if (flag1)
-            {
-                if (flag2)
-                {
+            if (flag1) {
+                if (flag2) {
                     Gui.drawRect(k1, i1 - 1, k1 + 1, (int) (i1 + 1 + RenderUtil.getStringHeight("_")), -3092272);
-                }
-                else
-                {
-                    RenderUtil.drawStringWithShadow("_", (float)k1, (float)i1, i);
+                } else {
+                    RenderUtil.drawStringWithShadow("_", (float) k1, (float) i1, i);
                 }
             }
 
-            if (k != j)
-            {
+            if (k != j) {
                 int l1 = (int) (l + RenderUtil.getStringWidth(s.substring(0, k)));
-                this.drawSelectionBox(k1, i1 - 1, l1 - 1, i1 + 1 + (int)RenderUtil.getStringHeight("_"));
+                this.drawSelectionBox(k1, i1 - 1, l1 - 1, i1 + 1 + (int) RenderUtil.getStringHeight("_"));
             }
         }
     }
 
-    private void drawSelectionBox(int startX, int startY, int endX, int endY)
-    {
-        if (startX < endX)
-        {
+    private void drawSelectionBox(int startX, int startY, int endX, int endY) {
+        if (startX < endX) {
             int i = startX;
             startX = endX;
             endX = i;
         }
 
-        if (startY < endY)
-        {
+        if (startY < endY) {
             int j = startY;
             startY = endY;
             endY = j;
         }
 
-        if (endX > this.x + this.width)
-        {
+        if (endX > this.x + this.width) {
             endX = this.x + this.width;
         }
 
-        if (startX > this.x + this.width)
-        {
+        if (startX > this.x + this.width) {
             startX = this.x + this.width;
         }
 
@@ -203,10 +167,10 @@ public class SalGuiTextField extends GuiTextField
         GlStateManager.enableColorLogic();
         GlStateManager.colorLogicOp(GlStateManager.LogicOp.OR_REVERSE);
         bufferbuilder.begin(7, DefaultVertexFormats.POSITION);
-        bufferbuilder.pos((double)startX, (double)endY, 0.0D).endVertex();
-        bufferbuilder.pos((double)endX, (double)endY, 0.0D).endVertex();
-        bufferbuilder.pos((double)endX, (double)startY, 0.0D).endVertex();
-        bufferbuilder.pos((double)startX, (double)startY, 0.0D).endVertex();
+        bufferbuilder.pos(startX, endY, 0.0D).endVertex();
+        bufferbuilder.pos(endX, endY, 0.0D).endVertex();
+        bufferbuilder.pos(endX, startY, 0.0D).endVertex();
+        bufferbuilder.pos(startX, startY, 0.0D).endVertex();
         tessellator.draw();
         GlStateManager.disableColorLogic();
         GlStateManager.enableTexture2D();
