+++ src/main/java/me/ionar/salhack/util/MathUtil.java
@@ -5,37 +5,32 @@ import net.minecraft.entity.Entity;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 
-import java.math.*;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 
-public final class MathUtil
-{
+public final class MathUtil {
 
     private static final int bigInt = 16 * 1024;
 
-    public static Vec3d interpolateEntity(Entity entity, float time)
-    {
+    public static Vec3d interpolateEntity(Entity entity, float time) {
         return new Vec3d(entity.lastTickPosX + (entity.posX - entity.lastTickPosX) * time,
                 entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * time,
                 entity.lastTickPosZ + (entity.posZ - entity.lastTickPosZ) * time);
     }
 
-    public static double radToDeg(double rad)
-    {
+    public static double radToDeg(double rad) {
         return rad * (float) (180.0f / Math.PI);
     }
 
-    public static double degToRad(double deg)
-    {
+    public static double degToRad(double deg) {
         return deg * (float) (Math.PI / 180.0f);
     }
 
-    public static Vec3d direction(float yaw)
-    {
+    public static Vec3d direction(float yaw) {
         return new Vec3d(Math.cos(degToRad(yaw + 90f)), 0, Math.sin(degToRad(yaw + 90f)));
     }
 
-    public static float[] calcAngle(Vec3d from, Vec3d to)
-    {
+    public static float[] calcAngle(Vec3d from, Vec3d to) {
         final double difX = to.x - from.x;
         final double difY = (to.y - from.y) * -1.0F;
         final double difZ = to.z - from.z;
@@ -43,37 +38,29 @@ public final class MathUtil
         final double dist = MathHelper.sqrt(difX * difX + difZ * difZ);
 
         return new float[]
-        { (float) MathHelper.wrapDegrees(Math.toDegrees(Math.atan2(difZ, difX)) - 90.0f),
-                (float) MathHelper.wrapDegrees(Math.toDegrees(Math.atan2(difY, dist))) };
+                {(float) MathHelper.wrapDegrees(Math.toDegrees(Math.atan2(difZ, difX)) - 90.0f),
+                        (float) MathHelper.wrapDegrees(Math.toDegrees(Math.atan2(difY, dist)))};
     }
 
-    public static double[] directionSpeed(double speed)
-    {
+    public static double[] directionSpeed(double speed) {
         final Minecraft mc = Minecraft.getMinecraft();
         float forward = mc.player.movementInput.moveForward;
         float side = mc.player.movementInput.moveStrafe;
         float yaw = mc.player.prevRotationYaw
                 + (mc.player.rotationYaw - mc.player.prevRotationYaw) * mc.getRenderPartialTicks();
 
-        if (forward != 0)
-        {
-            if (side > 0)
-            {
+        if (forward != 0) {
+            if (side > 0) {
                 yaw += (forward > 0 ? -45 : 45);
-            }
-            else if (side < 0)
-            {
+            } else if (side < 0) {
                 yaw += (forward > 0 ? 45 : -45);
             }
             side = 0;
 
             // forward = clamp(forward, 0, 1);
-            if (forward > 0)
-            {
+            if (forward > 0) {
                 forward = 1;
-            }
-            else if (forward < 0)
-            {
+            } else if (forward < 0) {
                 forward = -1;
             }
         }
@@ -83,40 +70,32 @@ public final class MathUtil
         final double posX = (forward * speed * cos + side * speed * sin);
         final double posZ = (forward * speed * sin - side * speed * cos);
         return new double[]
-        { posX, posZ };
+                {posX, posZ};
     }
 
-    public static double[] directionSpeedNoForward(double speed)
-    {
+    public static double[] directionSpeedNoForward(double speed) {
         final Minecraft mc = Minecraft.getMinecraft();
         float forward = 1f;
-        
+
         if (mc.gameSettings.keyBindLeft.isPressed() || mc.gameSettings.keyBindRight.isPressed() || mc.gameSettings.keyBindBack.isPressed() || mc.gameSettings.keyBindForward.isPressed())
             forward = mc.player.movementInput.moveForward;
-        
+
         float side = mc.player.movementInput.moveStrafe;
         float yaw = mc.player.prevRotationYaw
                 + (mc.player.rotationYaw - mc.player.prevRotationYaw) * mc.getRenderPartialTicks();
 
-        if (forward != 0)
-        {
-            if (side > 0)
-            {
+        if (forward != 0) {
+            if (side > 0) {
                 yaw += (forward > 0 ? -45 : 45);
-            }
-            else if (side < 0)
-            {
+            } else if (side < 0) {
                 yaw += (forward > 0 ? 45 : -45);
             }
             side = 0;
 
             // forward = clamp(forward, 0, 1);
-            if (forward > 0)
-            {
+            if (forward > 0) {
                 forward = 1;
-            }
-            else if (forward < 0)
-            {
+            } else if (forward < 0) {
                 forward = -1;
             }
         }
@@ -126,58 +105,48 @@ public final class MathUtil
         final double posX = (forward * speed * cos + side * speed * sin);
         final double posZ = (forward * speed * sin - side * speed * cos);
         return new double[]
-        { posX, posZ };
+                {posX, posZ};
     }
-    
-    public static Vec3d mult(Vec3d factor, Vec3d multiplier)
-    {
+
+    public static Vec3d mult(Vec3d factor, Vec3d multiplier) {
         return new Vec3d(factor.x * multiplier.x, factor.y * multiplier.y, factor.z * multiplier.z);
     }
 
-    public static Vec3d mult(Vec3d factor, float multiplier)
-    {
+    public static Vec3d mult(Vec3d factor, float multiplier) {
         return new Vec3d(factor.x * multiplier, factor.y * multiplier, factor.z * multiplier);
     }
 
-    public static Vec3d div(Vec3d factor, Vec3d divisor)
-    {
+    public static Vec3d div(Vec3d factor, Vec3d divisor) {
         return new Vec3d(factor.x / divisor.x, factor.y / divisor.y, factor.z / divisor.z);
     }
 
-    public static Vec3d div(Vec3d factor, float divisor)
-    {
+    public static Vec3d div(Vec3d factor, float divisor) {
         return new Vec3d(factor.x / divisor, factor.y / divisor, factor.z / divisor);
     }
 
-    public static double round(double value, int places)
-    {
-        if (places < 0)
-        {
+    public static double round(double value, int places) {
+        if (places < 0) {
             return value;
         }
         return new BigDecimal(value).setScale(places, RoundingMode.HALF_UP).doubleValue();
     }
 
     public static int ceil(float value) {
-        return bigInt - (int)((double)bigInt - value);
+        return bigInt - (int) ((double) bigInt - value);
     }
 
 
-    public static float clamp(float val, float min, float max)
-    {
-        if (val <= min)
-        {
+    public static float clamp(float val, float min, float max) {
+        if (val <= min) {
             val = min;
         }
-        if (val >= max)
-        {
+        if (val >= max) {
             val = max;
         }
         return val;
     }
 
-    public static float wrap(float val)
-    {
+    public static float wrap(float val) {
         val = val % 360.0f;
         if (val >= 180.0f)
             val -= 360.0f;
@@ -187,34 +156,29 @@ public final class MathUtil
     }
 
     // linearly maps value from the range (a..b) to (c..d)
-    public static double map(double value, double a, double b, double c, double d)
-    {
+    public static double map(double value, double a, double b, double c, double d) {
         // first map value from (a..b) to (0..1)
         value = (value - a) / (b - a);
         // then map it from (0..1) to (c..d) and return it
         return c + value * (d - c);
     }
 
-    public static double linear(double from, double to, double incline)
-    {
+    public static double linear(double from, double to, double incline) {
         return (from < to - incline) ? (from + incline) : ((from > to + incline) ? (from - incline) : to);
     }
 
-    public static double parabolic(double from, double to, double incline)
-    {
+    public static double parabolic(double from, double to, double incline) {
         return from + (to - from) / incline;
     }
 
-    public static double getDistance(Vec3d pos, double x, double y, double z)
-    {
+    public static double getDistance(Vec3d pos, double x, double y, double z) {
         final double deltaX = pos.x - x;
         final double deltaY = pos.y - y;
         final double deltaZ = pos.z - z;
-        return (double) MathHelper.sqrt(deltaX * deltaX + deltaY * deltaY + deltaZ * deltaZ);
+        return MathHelper.sqrt(deltaX * deltaX + deltaY * deltaY + deltaZ * deltaZ);
     }
 
-    public static double[] calcIntersection(double[] line, double[] line2)
-    {
+    public static double[] calcIntersection(double[] line, double[] line2) {
         final double a1 = line[3] - line[1];
         final double b1 = line[0] - line[2];
         final double c1 = a1 * line[0] + b1 * line[1];
@@ -226,14 +190,13 @@ public final class MathUtil
         final double delta = a1 * b2 - a2 * b1;
 
         return new double[]
-        { (b2 * c1 - b1 * c2) / delta, (a1 * c2 - a2 * c1) / delta };
+                {(b2 * c1 - b1 * c2) / delta, (a1 * c2 - a2 * c1) / delta};
     }
-    
-    public static double calculateAngle(double x1, double y1, double x2, double y2)
-    {
+
+    public static double calculateAngle(double x1, double y1, double x2, double y2) {
         double angle = Math.toDegrees(Math.atan2(x2 - x1, y2 - y1));
         // Keep angle between 0 and 360
-        angle = angle + Math.ceil( -angle / 360 ) * 360;
+        angle = angle + Math.ceil(-angle / 360) * 360;
 
         return angle;
     }
